1e7d3510-f519-465d-9731-73ed93d2fb50/pyproject.toml
[tool.poetry]
name = "repl_python3_beta_DiscordBot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.0"
Flask = "^2.0.1"
replit = "^3.2.4"
mysql-connector = "^2.2.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
